+++ internal/control-plane/bootstrap/options.go
@@ -1,0 +1,14 @@
+package bootstrap
+
+import (
+	"github.com/solo-io/gloo/pkg/bootstrap"
+)
+
+type Options struct {
+	bootstrap.Options
+	EnvoyOptions EnvoyOptions
+}
+
+type EnvoyOptions struct {
+	BindAddress string
+}
