+++ pkg/bootstrap/options.go
@@ -41,11 +41,6 @@ type Options struct {
 	FileStorageOptions   StorageOptions
 	FileOptions          FileOptions
 	VaultOptions         VaultOptions
-	EnvoyOptions         EnvoyOptions
-}
-
-type EnvoyOptions struct {
-	BindAddress string
 }
 
 type StorageOptions struct {
